Fix functions with multiple statements
There was a glitch behind the thinking of the ast_expression_chain. As it
turns out, ->next is used to chain together multiple statements or
expressions. I cannot be reused to implement a chained expression. Instead,
a new `rhs` member was added, and the `term` was renamed to `lhs`.

Also add a toString() method for hash, an iterator concept, and a tuple type
to support hash table iteration of items.