vm: Add locals concept
Function local variables are detected and assigned in a locals list. At
compile time, the names of the variables are compiled into a list of names,
and opcodes are generated to fetch or store values based on the name's index
in the names list. At runtime, the values are stored and fetched from a
values list from the index represented by the opcode argument. This
represents about a 25% increase in performance.