eval: Remove OP_STORE_ARG_LOCAL opcode
This was emitted by the compiler to unpack the arguments passed to the
interpreter when a function is interpreted. But since it is always done, it's
done before any other opcodes are interpreted, and the arguments are passed to
the interpreter in the order in which they are defined in the locals list, then
there's no need for the compiler to generate the opcodes and no need to
interpret the opcodes. Instead, the unpacking can always be done by the
interpreter.