compile: Fix issue where constants are incorrectly reused
Because 0 == "hello", because "hello", when coerced to an integer, has a
value of 0 and so is equal to zero, any string would poentially be
reused as zero in bytecode outputted by the compiler, and visa versa.

This addresses the issue by requiring that constants to be reused also
be of the exact same type as the previously-emitted constant in addition
to also being consider equal.