oo: Improve support for dotted operations
Fix dotted operations such as Class().method(), where the dot was
previously not handled correctly.

Move the super magic to the InstanceObject code

Add debug for parsing of classes and magic properties (this and super).

The class which a new class extends from can be the result of an
arbitrary expression (rather than a simple name lookup).