eval: Fix future-defined closure vars
If a variable is defined locally but is not assigned a value prior to
when a closure was created, previously the value would remain NULL in
the closure. This fixes the issue by moving the locals in the stack
frame onto the heap so that changes made to the variables in the closure
are reflected both in the stack frame and in all closures having access
to the variables.

Also, fix quite a bit of the bad spacing in the VM code.