eval: Optimize recursive calls
This adds a precursor to tail recursion. If a function appears to be
calling itself, optimize the VM in a way by removing the opcode to
lookup the function and add a new opcode to recursively call the
self-same code. This seems to have about a 10% impact on the fib.lox
test script.