vm: Add support to stash object instead of freeing
This creates the idea of a bucket where objects can be stored instead of
freeing and reallocating. This is very useful for simple objects like
Integer and Float where objects are generally considered immutable and
represent a simple value. Manipulations of them result in the creation
of new objects. Therefore, the VM spends a considerable amount of time
managing memory unnecessarily.

So instead of freeing simple objects, we stash them and reuse them. This
initial code chooses a relatively small stash size and does not handle
cases where the stash is filled up with objects of a certain type which
are not needed. In such a case the stash become unuseful. Eventually,
the stash should become a hashtable of stashes based on ObjectType which
will reduce contention and guarantee that objects of one type do not
fill up the stash.