Add first-class iterators
This adds a proper Iterator type and allows for better subtypes. It's
still pretty terrible, but there's a proof of concept with the
StringTree::chunks() method. What's next is to fully implement the
::iterate type function and add the iterative for loop to the compiler
and VM.

This also adds a "LoxStopIteration" type to be a special object separate
end of iteration from all other possible object types which might be
part of an iterated sequence.