Add table literal to the language
Allows for a table expression such as {'a': 1} and such. Any valid
expression is legal for both the keys and the values.

Python has an optimization to use a compressed version of opcodes by
creating a tuple of all the keys and using that as one constant value.
That way there is only the number of values plus one items pushed on the
stack for the operation. This is not implemented here but could serve a
great value of all the keys are constants.