eval: Use an func offset for binary math
Instead of using a nested switch block for the math operation, calculate
the location of the function based on the organization of the ObjectType
struct along with the opcode argument. If the opcode argument is an
index into the binary math functions in the ObjectType struct, then just
advancing ahead of `op_plus` by the pc->arg, skipping ahead by the size
of the function pointer, will arrive at the pointer to the operator
method. This includes a security check to ensure the pc->arg does not
overflow the expected functions.

This seems to have about a 5% impact for math-focused code such as the
fibbonacci sequence test (fib.lox).