compile: Fix rewrite of instr. by compile_assign
Instead of rewriting the output of compile_assign, pass the output
location of the evaluation to the expression compiler. This is a lot
more straight forward and fixes the unnecessarily allocated register
when the register is not required for the output of an expression (e.g.
if it is to be assigned to a local or global variable instead of auto
register).