opt: For attributes use pre-computed hashes
Generally speaking, the attribute names used in the bytecode are
constants (strings) which have precomputed hash values. However, the
storage for object attributes is a hash table. So using the pre-computed
hash value for the string property means that the hash of the string
constant will not need to be recomputed over and over again. This is
especially performant for attribute usage in loops.