builtin: Add `import` function
It's very rudimentary, but allows for the basis of an importing system.
Imported code is compiled, and the globals table from the eval process
is returned to the caller. This allows accessing functions, classes, and
properties of the module from the table.

A great next step would be to add a formal `import` keyword and allow
automated importing by the eval system.