assert: Add basic support for an assert statement
This will allow for basic assertions such as:

    assert(expression)
    assert(expression, "Error message")

For now, the second argument to the assertion is eagerly evaluated, and
so will be evaluated whether or not the asserted expression is true or
false.

An exception is raised, but the exception does not have a specific type
yet. Ultimately, an AssertionException or similar should be raised so
that it could be caught and handled somewhere.