repl: Fix crashes reading blocks
This fixes issues where reading blank lines would cause the interpreter
to crash.

Furthermore, it fixes issues where the repl code would not detect inner
lines within a block as being part of the block. So it would start to
evaluate the code inside the block rather than awaiting the end of the
block.